+++ netloom/__main__.py
@@ -1,4 +1,3 @@
 from netloom.cli.main import main
 
-
 main()